Satisfy repository size budgets
Problem:
The project tree ordering fix added regression coverage and catalog helpers to files that were already at their repository size budgets, causing the PR lint job to fail.

Root cause:
The behavior was correct, but the new tests and helpers were colocated with oversized legacy files instead of focused modules.

Fix:
Move catalog completeness and reconcile publication coverage into dedicated test files, relocate session-catalog helpers out of tabs.go, and remove a redundant reconcile comment without changing runtime behavior or lint baselines.

Verification:
- go run ./tools/repolint
- go test ./...
- cd desktop && go test ./...
- focused root and desktop go test -race
- cd desktop/frontend && pnpm test
- pnpm typecheck
- pnpm test:typecheck
- pnpm lint:hooks
